requirements.txt
flask
flask-cors
gunicorn
requests
numpy
pandas
torch
transformers
scikit-learn
openai
huggingface_hub
flask-socketio
flask-wtf
flask-login
flask-mail
flask-sqlalchemy
flask-migrate
flask-restful
flask-jwt-extended
sqlalchemy
pymysql
bcrypt
itsdangerous
jinja2
google-generativeai
